+++ src/containerapp/azext_containerapp/_help.py
@@ -2135,9 +2135,9 @@ helps['containerapp job registry set'] = """
 # Maintenance Config Commands
 helps['containerapp env maintenance-config add'] = """
     type: command
-    short-summary: Add a Maintenance Configuration to a Container App Environment
+    short-summary: Add Planned Maintenance to a Container App Environment
     examples:
-    - name: Configure an Env to use a maintenance config
+    - name: Configure a Container App Environment to use a Planned Maintenance
       text: |
           az containerapp env maintenance-config add --env-name myEnv -g MyResourceGroup \\
               --duration 10 --start-hour-utc 11 --weekday Sunday
@@ -2145,9 +2145,9 @@ helps['containerapp env maintenance-config add'] = """
 
 helps['containerapp env maintenance-config update'] = """
     type: command
-    short-summary: Update a Maintenance Configuration in a Container App Environment
+    short-summary: Update Planned Maintenance in a Container App Environment
     examples:
-    - name: Configure an Env to use a maintenance config
+    - name: Update the Planned Maintenance in a Container App Environment
       text: |
           az containerapp env maintenance-config update --env-name myEnv -g MyResourceGroup \\
               --duration 8 --start-hour-utc 12 --weekday Thursday
@@ -2155,18 +2155,18 @@ helps['containerapp env maintenance-config update'] = """
 
 helps['containerapp env maintenance-config show'] = """
     type: command
-    short-summary: Show a Maintenance Configuration in a Container App Environment
+    short-summary: Show Planned Maintenance in a Container App Environment
     examples:
-    - name: Show a maintenance config
+    - name: Show Planned Maintenance
       text: |
           az containerapp env maintenance-config show --env-name myEnv -g MyResourceGroup
 """
 
 helps['containerapp env maintenance-config delete'] = """
     type: command
-    short-summary: Delete a Maintenance Configuration in a Container App Environment
+    short-summary: Delete Planned Maintenance in a Container App Environment
     examples:
-    - name: Delete a Maintenance Config
+    - name: Delete Planned Maintenance
       text: |
-          az containerapp env maintenance-config add --env-name myEnv -g MyResourceGroup
+          az containerapp env maintenance-config delete --env-name myEnv -g MyResourceGroup
 """